moved components and fixed loading spinner
