+++ env-config/example.env
@@ -1,5 +1,19 @@
 SENTRY_ORG=
 
+NEXT_PUBLIC_DSN=
+NEXT_PUBLIC_ENVIRONMENT=
+NEXT_RELEASE_PACKAGE_NAME=
+NEXT_SENTRY_PROJECT=
+NEXT_SOURCEMAPS_DIR=
+NEXT_SOURCEMAPS_URL_PREFIX=
+NEXT_PUBLIC_FLASK_BACKEND=
+NEXT_PUBLIC_EXPRESS_BACKEND=
+NEXT_PUBLIC_SPRINGBOOT_BACKEND=
+NEXT_PUBLIC_ASPNETCORE_BACKEND=
+NEXT_PUBLIC_LARAVEL_BACKEND=
+NEXT_PUBLIC_RUBY_BACKEND=
+NEXT_PUBLIC_RUBYONRAILS_BACKEND=
+
 REACT_APP_DSN=
 REACT_APP_ENGINE_SERVICE=
 REACT_APP_ENVIRONMENT=
