+++ next/package.json
@@ -16,7 +16,7 @@
     "prisma": "^5.21.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-loader-spinner": "~4.0.0",
+    "react-loader-spinner": "~6.1.6",
     "react-redux": "~7.2.4",
     "react-router-dom": "~6.3.0",
     "react-scripts": "~5.0.1",
