+++ next/src/app/page.jsx
@@ -18,9 +18,8 @@ const divStyle = {
 
 
 export default function Page() {
-  console.log("in home page");
   const router = useRouter();
-  const { backend, frontendSlowdown } = useSearchParams();
+  const { backend } = useSearchParams();
   const backendType = determineBackendType(backend);
   const backendUrl = determineBackendUrl(backendType);
   console.log('backend is ' + backendUrl);
