+++ next/src/components/SentryQueryInitializer.jsx
