+++ next/src/components/ThreeDotLoader.jsx
@@ -1,11 +1,13 @@
 'use client'
 
-import Loader from 'react-loader-spinner';
+import { ThreeDots } from 'react-loader-spinner';
+
 
 export default function ThreeDotLoader() {
+
   return (
     <div className="loader-container">
-      <Loader type="ThreeDots" color="#f6cfb2" height={150} width={150} />
+      <ThreeDots color="#f6cfb2" height={150} width={150} />
     </div>
   );
 }
