+++ next/src/styles/products.css
@@ -36,4 +36,6 @@
 
 .loader-container {
   text-align: center;
+  display: flex;
+  justify-content: center;
 }
