--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "react-loader-spinner": "^6.1.6"
-  }
-}
